+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.2.0 (git develop)
 
+- references_many.first() will only find first object from DB if proxy has not been loaded, or else return first element in loaded target.
 - Wrapper::Thrift has been moved into Adapter::Thrift. Adding more adapters should be not that hard now.
 - References many is now possible. We have to strategies: Store an array of foreign keys in the owner,
   or supply a ids-starts-with and open up a scanner and read from that point.